+++ pyshbundle/conv_sh.py
