+++ pyshbundle/io.py
