+++ pyshbundle/pysh_core.py
@@ -10,7 +10,7 @@ from scipy import linalg
 
 from os import chdir, getcwd
 
-from pyshbundle.convert_sh_fmt import cs2sc, sc2cs
+from pyshbundle.conv_sh import cs2sc, sc2cs
 from pyshbundle.shutils import normalklm, PLM, iplm, eigengrav, ispec, Gaussian, naninterp, neumann
